Move loggregator util function to metrics module
